refactor(ui): 重构 UI 组件库
- 将 UI 组件库的代码结构进行重新组织，优化目录结构
- 更新 package.json 文件，调整依赖项和配置
- 重命名部分文件和目录，提高代码的可维护性
- 更新 index.ts 文件，改进组件库的导出方式
