+++ packages/ui/src/components/svg-icon/index.ts
