+++ packages/ui/src/components/svg-icon/svg-icon.vue
